Cheater Detection & Caching
Added a rudimentary cache, that stores some player data at a tickcount.
Added aimbot detection to cheater detection.
Included region selector files in project.
Added a check to make sure we were up to date with the server when trying to check players.
Added a "mark reduction" system for players that are behaving well.